fix(angular): fix 6 compiler divergences from Angular reference (#31)
1. viewport({trigger: ...}) now keeps empty options object (was dropping it)
2. Timer delays preserve fractional precision (u32 -> f64, e.g. 1.5ms)
3. Pipe detection in @for track now traverses all AST expression types
4. hover/interaction triggers enforce max 1 parameter; never removed from on-triggers
5. @for let assignment a=b=c correctly rejected (split instead of splitn)
6. @for expression parse failure no longer emits extra missing-track error

+++ crates/oxc_angular_compiler/src/ast/r3.rs
@@ -1084,8 +1084,8 @@ pub struct R3HoverDeferredTrigger<'a> {
 /// A timer deferred trigger.
 #[derive(Debug)]
 pub struct R3TimerDeferredTrigger {
-    /// Delay in milliseconds.
-    pub delay: u32,
+    /// Delay in milliseconds (f64 to preserve fractional precision).
+    pub delay: f64,
     /// Source span.
     pub source_span: Span,
     /// Name span.
